Fluidify performance increase
